refactor(workflows): extract step-format + template-list from index (phase A)
First, pure, zero-behavior-change slice of the workflows/index.ts split.
Extracts two groups of module-scope helpers that have no pluginCtx
coupling and no module-load concerns:

- lib/step-format.ts   — formatSchema + formatStepContext (pure
                         presentation-for-agents string formatting)
- lib/template-list.ts — resolveSubWorkflows, buildTemplateList,
                         workflowSkillDriftBySkill/ForDefinition,
                         collectWorkflowSkillRefs, countSteps (pure
                         definition + skill-drift aggregation)

index.ts drops from 2,146 to ~1,950 and imports both back. The riskier
route/search/validation/gate extractions are inseparable from deduping
index's triple-scope helper copies and replacing the mutable pluginCtx
module global with ctx injection (behavior-touching), so they are tiered
into follow-up PRs — see tasks/plan.md.

Verified: typecheck + eslint + workflows suite (419/0) + full suite
(5123/0) + binary build + isolated-home boot smoke (Workflows Ready;
GET /definitions [buildTemplateList path] and /node-types return 200).
